+++ rascal-vscode-extension/src/lsp/RascalLSPConnection.ts
@@ -33,157 +33,7 @@ import { integer, LanguageClient, LanguageClientOptions, ServerOptions, StreamIn
 import { getJavaExecutable } from '../auto-jvm/JavaLookup';
 import { RascalFileSystemProvider } from '../fs/RascalFileSystemProviders';
 import { VSCodeUriResolverServer } from '../fs/VSCodeURIResolver';
-
-enum LogLevel {
-    fatal = "FATAL",
-    error = "ERROR",
-    warn = "WARN",
-    info = "INFO",
-    debug = "DEBUG",
-    trace = "TRACE",
-}
-
-class JsonLogMessage {
-    readonly timestamp: Date;
-    readonly loglevel: LogLevel;
-    readonly message: string;
-    readonly threadName: string;
-    readonly loggerName: string;
-
-    private static readonly timestampKey = "@timestamp";
-    private static readonly loglevelKey = "log.level";
-    private static readonly messageKey = "message";
-    private static readonly threadNameKey = "process.thread.name";
-    private static readonly loggerNameKey = "log.logger";
-
-    constructor(timestamp: Date, loglevel: LogLevel, message: string, threadName: string, loggerName: string) {
-        this.timestamp = timestamp;
-        this.loglevel = loglevel;
-        this.message = message;
-        this.threadName = threadName;
-        this.loggerName = loggerName;
-    }
-
-    static parse(str: string): JsonLogMessage {
-        const obj = JSON.parse(str);
-        const timestamp: Date = new Date(obj[JsonLogMessage.timestampKey]);
-        const loglevel: LogLevel = obj[JsonLogMessage.loglevelKey];
-        const message: string = obj[JsonLogMessage.messageKey];
-        const threadName: string = obj[JsonLogMessage.threadNameKey];
-        const loggerName: string = obj[JsonLogMessage.loggerNameKey];
-
-        return new JsonLogMessage(timestamp, loglevel, message, threadName, loggerName);
-    }
-}
-
-class JsonParserOutputChannel implements vscode.OutputChannel {
-    readonly title: string;
-
-    private readonly nested: vscode.LogOutputChannel;
-
-    constructor(name: string) {
-        this.nested = vscode.window.createOutputChannel(name, {log: true});
-        this.title = name;
-    }
-
-    private printLogOutput(loglevel: LogLevel, message: string) {
-        switch (loglevel) {
-            case LogLevel.fatal: // intentional fall-trough
-            case LogLevel.error: {
-                this.nested.error(message);
-                break;
-            }
-            case LogLevel.warn: {
-                this.nested.warn(message);
-                break;
-            }
-            case LogLevel.info: {
-                this.nested.info(message);
-                break;
-            }
-            case LogLevel.debug: {
-                this.nested.debug(message);
-                break;
-            }
-            case LogLevel.trace: {
-                this.nested.trace(message);
-                break;
-            }
-            default: {
-                this.nested.appendLine(`${loglevel} ${message}`);
-            }
-        }
-    }
-
-    private printPayload(payload: string): void {
-        for (const line of payload.trim().split("\n")) {
-            try {
-                this.printJsonPayLoad(line);
-            } catch (e) {
-                if (e instanceof SyntaxError) {
-                    // this was not JSON at all
-                    this.printNonJsonPayload(line);
-                } else {
-                    this.nested.appendLine(`Error while logging ${line}: ${e}`);
-                    throw e;
-                }
-            }
-        }
-    }
-
-    private printNonJsonPayload(payload: string): void {
-        // try to find the log level and message
-        const jsDefault = /^[^[]*\[(?<level>\w+) [^\]]*\]\s*(?<message>.*)$/si;
-        const log4jDefault = /^\s*[^\s]+\s+(?<thread>\w+)\s+(?<level>\w+)\s+(?<message>.*)$/si;
-
-        for (const pattern of [jsDefault, log4jDefault]) {
-            const matches = pattern.exec(payload);
-            if (matches?.groups && "level" in matches.groups && "message" in matches.groups) {
-                const logLevel = matches.groups["level"].toLocaleUpperCase() as LogLevel;
-                let message = matches.groups["message"];
-                if ("thread" in matches.groups) {
-                    message = `[${matches.groups["thread"]}] ${message}`;
-                }
-                this.printLogOutput(logLevel, message);
-                return;
-            }
-        }
-
-        // we do not understand the structure of this log message; just print it as-is
-        this.nested.appendLine(payload);
-    }
-
-    private printJsonPayLoad(payload: string): void {
-        const json: JsonLogMessage = JsonLogMessage.parse(payload);
-        // no timestamp or log level, since LogOutputChannel functions add those
-        const log = `[${json.threadName}] ${json.loggerName} ${json.message}`;
-        this.printLogOutput(json.loglevel, log);
-    }
-
-    append(payload: string): void {
-        this.printPayload(payload);
-    }
-
-    appendLine(payload: string): void {
-        this.printPayload(payload);
-    }
-
-    show(preserveFocus?: unknown): void {
-        this.nested.show(preserveFocus as boolean);
-    }
-    replace(value: string): void {
-        this.nested.replace(value);
-    }
-    clear(): void {
-        this.nested.clear();
-    }
-    hide(): void {
-        this.nested.hide();
-    }
-    dispose(): void {
-        this.nested.dispose();
-    }
-}
+import { JsonParserOutputChannel } from './JsonOutputChannel';
 
 export async function activateLanguageClient(
     { language, title, jarPath, vfsServer, isParametricServer = false, deployMode = true, devPort = -1, dedicated = false, lspArg = "" } :
